requirements.txt
pyyaml
future
opencv-python